requirements.txt
selenium==4.23.1
python-dotenv==1.0.1
discord==2.3.2
discord.py==2.4.0